factor more methods out of utils
