+++ sbb_newspapers_org_image/eynollah.py
@@ -39,6 +39,7 @@ from .utils.contour import (
     contours_in_same_horizon,
     filter_contours_area_of_image_interiors,
     filter_contours_area_of_image_tables,
+    filter_contours_area_of_image,
     find_contours_mean_y_diff,
     find_features_of_contours,
     find_new_features_of_contoures,
@@ -74,10 +75,14 @@ from .utils.separate_lines import (
     seperate_lines_vertical_cont,
 )
 
+from .utils.drop_capitals import (
+    adhere_drop_capital_region_into_cprresponding_textline,
+    filter_small_drop_capitals_from_no_patch_layout
+)
+
 from .utils import (
     boosting_headers_by_longshot_region_segmentation,
     crop_image_inside_box,
-    filter_small_drop_capitals_from_no_patch_layout,
     find_features_of_lines,
     find_num_col,
     find_num_col_by_vertical_lines,
@@ -2788,10 +2793,10 @@ class eynollah:
 
         median_blur=False
 
-        img= self.resize_image(img_org, int(img_org.shape[0]*ratio_y), int(img_org.shape[1]*ratio_x))
+        img= resize_image(img_org, int(img_org.shape[0]*ratio_y), int(img_org.shape[1]*ratio_x))
 
         if binary:
-            img = self.otsu_copy_binary(img)#self.otsu_copy(img)
+            img = otsu_copy_binary(img)#self.otsu_copy(img)
             img = img.astype(np.uint16)
 
         if median_blur:
@@ -2801,7 +2806,7 @@ class eynollah:
             img = img.astype(np.uint16)
         prediction_regions_org_y=self.do_prediction(patches,img,model_region)
 
-        prediction_regions_org_y=self.resize_image(prediction_regions_org_y, img_height_h, img_width_h )
+        prediction_regions_org_y = resize_image(prediction_regions_org_y, img_height_h, img_width_h )
 
         #plt.imshow(prediction_regions_org_y[:,:,0])
         #plt.show()
@@ -2823,10 +2828,10 @@ class eynollah:
         ratio_y=1
         median_blur=False
 
-        img= self.resize_image(img_org, int(img_org.shape[0]*ratio_y), int(img_org.shape[1]*ratio_x))
+        img= resize_image(img_org, int(img_org.shape[0]*ratio_y), int(img_org.shape[1]*ratio_x))
 
         if binary:
-            img = self.otsu_copy_binary(img)#self.otsu_copy(img)
+            img = otsu_copy_binary(img)#self.otsu_copy(img)
             img = img.astype(np.uint16)
 
         if median_blur:
@@ -2836,7 +2841,7 @@ class eynollah:
             img = img.astype(np.uint16)
         prediction_regions_org=self.do_prediction(patches,img,model_region)
 
-        prediction_regions_org=self.resize_image(prediction_regions_org, img_height_h, img_width_h )
+        prediction_regions_org=resize_image(prediction_regions_org, img_height_h, img_width_h )
 
         ##plt.imshow(prediction_regions_org[:,:,0])
         ##plt.show()
@@ -2863,10 +2868,10 @@ class eynollah:
         ratio_y=1
         median_blur=False
 
-        img= self.resize_image(img_org, int(img_org.shape[0]*ratio_y), int(img_org.shape[1]*ratio_x))
+        img= resize_image(img_org, int(img_org.shape[0]*ratio_y), int(img_org.shape[1]*ratio_x))
 
         if binary:
-            img = self.otsu_copy_binary(img)#self.otsu_copy(img)
+            img = otsu_copy_binary(img)#self.otsu_copy(img)
             img = img.astype(np.uint16)
 
         if median_blur:
@@ -2878,7 +2883,7 @@ class eynollah:
         marginal_patch=0.2
         prediction_regions_org2=self.do_prediction(patches,img,model_region,marginal_patch)
 
-        prediction_regions_org2=self.resize_image(prediction_regions_org2, img_height_h, img_width_h )
+        prediction_regions_org2=resize_image(prediction_regions_org2, img_height_h, img_width_h )
 
         #plt.imshow(prediction_regions_org2[:,:,0])
         #plt.show()
@@ -2998,11 +3003,11 @@ class eynollah:
 
         pixel_img=1
         min_area_text=0.00001
-        polygons_of_only_texts=self.return_contours_of_interested_region(mask_texts_only,pixel_img,min_area_text)
+        polygons_of_only_texts=return_contours_of_interested_region(mask_texts_only,pixel_img,min_area_text)
 
-        polygons_of_only_images=self.return_contours_of_interested_region(mask_images_only,pixel_img)
+        polygons_of_only_images=return_contours_of_interested_region(mask_images_only,pixel_img)
 
-        polygons_of_only_lines=self.return_contours_of_interested_region(mask_lines_only,pixel_img,min_area_text)
+        polygons_of_only_lines=return_contours_of_interested_region(mask_lines_only,pixel_img,min_area_text)
 
 
         text_regions_p_true=np.zeros(prediction_regions_org.shape)
@@ -3065,13 +3070,13 @@ class eynollah:
         if text_with_lines.shape[0]<=1500:
             pass
         elif text_with_lines.shape[0]>1500 and text_with_lines.shape[0]<=1800:
-            text_with_lines=self.resize_image(text_with_lines,int(text_with_lines.shape[0]*1.5),text_with_lines.shape[1])
+            text_with_lines=resize_image(text_with_lines,int(text_with_lines.shape[0]*1.5),text_with_lines.shape[1])
             text_with_lines=cv2.erode(text_with_lines,self.kernel,iterations=5)
-            text_with_lines=self.resize_image(text_with_lines,text_with_lines_eroded.shape[0],text_with_lines_eroded.shape[1])
+            text_with_lines=resize_image(text_with_lines,text_with_lines_eroded.shape[0],text_with_lines_eroded.shape[1])
         else:
-            text_with_lines=self.resize_image(text_with_lines,int(text_with_lines.shape[0]*1.8),text_with_lines.shape[1])
+            text_with_lines=resize_image(text_with_lines,int(text_with_lines.shape[0]*1.8),text_with_lines.shape[1])
             text_with_lines=cv2.erode(text_with_lines,self.kernel,iterations=7)
-            text_with_lines=self.resize_image(text_with_lines,text_with_lines_eroded.shape[0],text_with_lines_eroded.shape[1])
+            text_with_lines=resize_image(text_with_lines,text_with_lines_eroded.shape[0],text_with_lines_eroded.shape[1])
 
 
         text_with_lines_y=text_with_lines.sum(axis=0)
@@ -3193,7 +3198,7 @@ class eynollah:
                 mask_marginals[:,:]=1
 
             #print(mask_marginals.shape,point_left,point_right,'nadosh')
-            mask_marginals_rotated=self.rotate_image(mask_marginals,-slope_deskew)
+            mask_marginals_rotated=rotate_image(mask_marginals,-slope_deskew)
 
             #print(mask_marginals_rotated.shape,'nadosh')
             mask_marginals_rotated_sum=mask_marginals_rotated.sum(axis=0)
@@ -3227,9 +3232,9 @@ class eynollah:
 
             pixel_img=4
             min_area_text=0.00001
-            polygons_of_marginals=self.return_contours_of_interested_region(text_regions,pixel_img,min_area_text)
+            polygons_of_marginals=return_contours_of_interested_region(text_regions,pixel_img,min_area_text)
 
-            cx_text_only,cy_text_only ,x_min_text_only,x_max_text_only, y_min_text_only ,y_max_text_only,y_cor_x_min_main=self.find_new_features_of_contoures(polygons_of_marginals)
+            cx_text_only,cy_text_only ,x_min_text_only,x_max_text_only, y_min_text_only ,y_max_text_only,y_cor_x_min_main=find_new_features_of_contoures(polygons_of_marginals)
 
             text_regions[(text_regions[:,:]==4)]=1
 
@@ -3577,452 +3582,6 @@ class eynollah:
                     order_text_new.append(tartib_new)
 
             return order_text_new, id_of_texts_tot
-
-    def adhere_drop_capital_region_into_cprresponding_textline(self, text_regions_p, polygons_of_drop_capitals, contours_only_text_parent, contours_only_text_parent_h, all_box_coord, all_box_coord_h, all_found_texline_polygons, all_found_texline_polygons_h):
-        # print(np.shape(all_found_texline_polygons),np.shape(all_found_texline_polygons[3]),'all_found_texline_polygonsshape')
-        # print(all_found_texline_polygons[3])
-        cx_m, cy_m, _, _, _, _, _ = find_new_features_of_contoures(contours_only_text_parent)
-        cx_h, cy_h, _, _, _, _, _ = find_new_features_of_contoures(contours_only_text_parent_h)
-        cx_d, cy_d, _, _, y_min_d, y_max_d, _ = find_new_features_of_contoures(polygons_of_drop_capitals)
-
-        img_con_all = np.zeros((text_regions_p.shape[0], text_regions_p.shape[1], 3))
-        for j_cont in range(len(contours_only_text_parent)):
-            img_con_all[all_box_coord[j_cont][0] : all_box_coord[j_cont][1], all_box_coord[j_cont][2] : all_box_coord[j_cont][3], 0] = (j_cont + 1) * 3
-            # img_con_all=cv2.fillPoly(img_con_all,pts=[contours_only_text_parent[j_cont]],color=((j_cont+1)*3,(j_cont+1)*3,(j_cont+1)*3))
-
-        # plt.imshow(img_con_all[:,:,0])
-        # plt.show()
-        # img_con_all=cv2.dilate(img_con_all, self.kernel, iterations=3)
-
-        # plt.imshow(img_con_all[:,:,0])
-        # plt.show()
-        # print(np.unique(img_con_all[:,:,0]))
-        for i_drop in range(len(polygons_of_drop_capitals)):
-            # print(i_drop,'i_drop')
-            img_con_all_copy = np.copy(img_con_all)
-            img_con = np.zeros((text_regions_p.shape[0], text_regions_p.shape[1], 3))
-            img_con = cv2.fillPoly(img_con, pts=[polygons_of_drop_capitals[i_drop]], color=(1, 1, 1))
-
-            # plt.imshow(img_con[:,:,0])
-            # plt.show()
-            ##img_con=cv2.dilate(img_con, self.kernel, iterations=30)
-
-            # plt.imshow(img_con[:,:,0])
-            # plt.show()
-
-            # print(np.unique(img_con[:,:,0]))
-
-            img_con_all_copy[:, :, 0] = img_con_all_copy[:, :, 0] + img_con[:, :, 0]
-
-            img_con_all_copy[:, :, 0][img_con_all_copy[:, :, 0] == 1] = 0
-
-            kherej_ghesmat = np.unique(img_con_all_copy[:, :, 0]) / 3
-            res_summed_pixels = np.unique(img_con_all_copy[:, :, 0]) % 3
-            region_with_intersected_drop = kherej_ghesmat[res_summed_pixels == 1]
-            # region_with_intersected_drop=region_with_intersected_drop/3
-            region_with_intersected_drop = region_with_intersected_drop.astype(np.uint8)
-
-            # print(len(region_with_intersected_drop),'region_with_intersected_drop1')
-            if len(region_with_intersected_drop) == 0:
-                img_con_all_copy = np.copy(img_con_all)
-                img_con = cv2.dilate(img_con, self.kernel, iterations=4)
-
-                img_con_all_copy[:, :, 0] = img_con_all_copy[:, :, 0] + img_con[:, :, 0]
-
-                img_con_all_copy[:, :, 0][img_con_all_copy[:, :, 0] == 1] = 0
-
-                kherej_ghesmat = np.unique(img_con_all_copy[:, :, 0]) / 3
-                res_summed_pixels = np.unique(img_con_all_copy[:, :, 0]) % 3
-                region_with_intersected_drop = kherej_ghesmat[res_summed_pixels == 1]
-                # region_with_intersected_drop=region_with_intersected_drop/3
-                region_with_intersected_drop = region_with_intersected_drop.astype(np.uint8)
-            # print(np.unique(img_con_all_copy[:,:,0]))
-            if self.curved_line:
-
-                if len(region_with_intersected_drop) > 1:
-                    sum_pixels_of_intersection = []
-                    for i in range(len(region_with_intersected_drop)):
-                        # print((region_with_intersected_drop[i]*3+1))
-                        sum_pixels_of_intersection.append(((img_con_all_copy[:, :, 0] == (region_with_intersected_drop[i] * 3 + 1)) * 1).sum())
-                    # print(sum_pixels_of_intersection)
-                    region_final = region_with_intersected_drop[np.argmax(sum_pixels_of_intersection)] - 1
-
-                    # print(region_final,'region_final')
-                    # cx_t,cy_t ,_, _, _ ,_,_= find_new_features_of_contoures(all_found_texline_polygons[int(region_final)])
-                    try:
-                        cx_t, cy_t, _, _, _, _, _ = find_new_features_of_contoures(all_found_texline_polygons[int(region_final)])
-                        # print(all_box_coord[j_cont])
-                        # print(cx_t)
-                        # print(cy_t)
-                        # print(cx_d[i_drop])
-                        # print(cy_d[i_drop])
-                        y_lines = np.array(cy_t)  # all_box_coord[int(region_final)][0]+np.array(cy_t)
-
-                        # print(y_lines)
-
-                        y_lines[y_lines < y_min_d[i_drop]] = 0
-                        # print(y_lines)
-
-                        arg_min = np.argmin(np.abs(y_lines - y_min_d[i_drop]))
-                        # print(arg_min)
-
-                        cnt_nearest = np.copy(all_found_texline_polygons[int(region_final)][arg_min])
-                        cnt_nearest[:, 0, 0] = all_found_texline_polygons[int(region_final)][arg_min][:, 0, 0]  # +all_box_coord[int(region_final)][2]
-                        cnt_nearest[:, 0, 1] = all_found_texline_polygons[int(region_final)][arg_min][:, 0, 1]  # +all_box_coord[int(region_final)][0]
-
-                        img_textlines = np.zeros((text_regions_p.shape[0], text_regions_p.shape[1], 3))
-                        img_textlines = cv2.fillPoly(img_textlines, pts=[cnt_nearest], color=(255, 255, 255))
-                        img_textlines = cv2.fillPoly(img_textlines, pts=[polygons_of_drop_capitals[i_drop]], color=(255, 255, 255))
-
-                        img_textlines = img_textlines.astype(np.uint8)
-                        imgray = cv2.cvtColor(img_textlines, cv2.COLOR_BGR2GRAY)
-                        ret, thresh = cv2.threshold(imgray, 0, 255, 0)
-
-                        contours_combined, hierachy = cv2.findContours(thresh, cv2.RETR_TREE, cv2.CHAIN_APPROX_SIMPLE)
-
-                        # print(len(contours_combined),'len textlines mixed')
-                        areas_cnt_text = np.array([cv2.contourArea(contours_combined[j]) for j in range(len(contours_combined))])
-
-                        contours_biggest = contours_combined[np.argmax(areas_cnt_text)]
-
-                        # print(np.shape(contours_biggest))
-                        # print(contours_biggest[:])
-                        # contours_biggest[:,0,0]=contours_biggest[:,0,0]#-all_box_coord[int(region_final)][2]
-                        # contours_biggest[:,0,1]=contours_biggest[:,0,1]#-all_box_coord[int(region_final)][0]
-
-                        # contours_biggest=contours_biggest.reshape(np.shape(contours_biggest)[0],np.shape(contours_biggest)[2])
-
-                        all_found_texline_polygons[int(region_final)][arg_min] = contours_biggest
-
-                    except:
-                        # print('gordun1')
-                        pass
-                elif len(region_with_intersected_drop) == 1:
-                    region_final = region_with_intersected_drop[0] - 1
-
-                    # areas_main=np.array([cv2.contourArea(all_found_texline_polygons[int(region_final)][0][j] ) for j in range(len(all_found_texline_polygons[int(region_final)]))])
-
-                    # cx_t,cy_t ,_, _, _ ,_,_= find_new_features_of_contoures(all_found_texline_polygons[int(region_final)])
-
-                    cx_t, cy_t, _, _, _, _, _ = find_new_features_of_contoures(all_found_texline_polygons[int(region_final)])
-                    # print(all_box_coord[j_cont])
-                    # print(cx_t)
-                    # print(cy_t)
-                    # print(cx_d[i_drop])
-                    # print(cy_d[i_drop])
-                    y_lines = np.array(cy_t)  # all_box_coord[int(region_final)][0]+np.array(cy_t)
-
-                    y_lines[y_lines < y_min_d[i_drop]] = 0
-                    # print(y_lines)
-
-                    arg_min = np.argmin(np.abs(y_lines - y_min_d[i_drop]))
-                    # print(arg_min)
-
-                    cnt_nearest = np.copy(all_found_texline_polygons[int(region_final)][arg_min])
-                    cnt_nearest[:, 0, 0] = all_found_texline_polygons[int(region_final)][arg_min][:, 0, 0]  # +all_box_coord[int(region_final)][2]
-                    cnt_nearest[:, 0, 1] = all_found_texline_polygons[int(region_final)][arg_min][:, 0, 1]  # +all_box_coord[int(region_final)][0]
-
-                    img_textlines = np.zeros((text_regions_p.shape[0], text_regions_p.shape[1], 3))
-                    img_textlines = cv2.fillPoly(img_textlines, pts=[cnt_nearest], color=(255, 255, 255))
-                    img_textlines = cv2.fillPoly(img_textlines, pts=[polygons_of_drop_capitals[i_drop]], color=(255, 255, 255))
-
-                    img_textlines = img_textlines.astype(np.uint8)
-
-                    # plt.imshow(img_textlines)
-                    # plt.show()
-                    imgray = cv2.cvtColor(img_textlines, cv2.COLOR_BGR2GRAY)
-                    ret, thresh = cv2.threshold(imgray, 0, 255, 0)
-
-                    contours_combined, hierachy = cv2.findContours(thresh, cv2.RETR_TREE, cv2.CHAIN_APPROX_SIMPLE)
-
-                    # print(len(contours_combined),'len textlines mixed')
-                    areas_cnt_text = np.array([cv2.contourArea(contours_combined[j]) for j in range(len(contours_combined))])
-
-                    contours_biggest = contours_combined[np.argmax(areas_cnt_text)]
-
-                    # print(np.shape(contours_biggest))
-                    # print(contours_biggest[:])
-                    # contours_biggest[:,0,0]=contours_biggest[:,0,0]#-all_box_coord[int(region_final)][2]
-                    # contours_biggest[:,0,1]=contours_biggest[:,0,1]#-all_box_coord[int(region_final)][0]
-                    # print(np.shape(contours_biggest),'contours_biggest')
-                    # print(np.shape(all_found_texline_polygons[int(region_final)][arg_min]))
-                    ##contours_biggest=contours_biggest.reshape(np.shape(contours_biggest)[0],np.shape(contours_biggest)[2])
-                    all_found_texline_polygons[int(region_final)][arg_min] = contours_biggest
-
-                    # print(cx_t,'print')
-                    try:
-                        # print(all_found_texline_polygons[j_cont][0])
-                        cx_t, cy_t, _, _, _, _, _ = find_new_features_of_contoures(all_found_texline_polygons[int(region_final)])
-                        # print(all_box_coord[j_cont])
-                        # print(cx_t)
-                        # print(cy_t)
-                        # print(cx_d[i_drop])
-                        # print(cy_d[i_drop])
-                        y_lines = all_box_coord[int(region_final)][0] + np.array(cy_t)
-
-                        y_lines[y_lines < y_min_d[i_drop]] = 0
-                        # print(y_lines)
-
-                        arg_min = np.argmin(np.abs(y_lines - y_min_d[i_drop]))
-                        # print(arg_min)
-
-                        cnt_nearest = np.copy(all_found_texline_polygons[int(region_final)][arg_min])
-                        cnt_nearest[:, 0, 0] = all_found_texline_polygons[int(region_final)][arg_min][:, 0, 0]  # +all_box_coord[int(region_final)][2]
-                        cnt_nearest[:, 0, 1] = all_found_texline_polygons[int(region_final)][arg_min][:, 0, 1]  # +all_box_coord[int(region_final)][0]
-
-                        img_textlines = np.zeros((text_regions_p.shape[0], text_regions_p.shape[1], 3))
-                        img_textlines = cv2.fillPoly(img_textlines, pts=[cnt_nearest], color=(255, 255, 255))
-                        img_textlines = cv2.fillPoly(img_textlines, pts=[polygons_of_drop_capitals[i_drop]], color=(255, 255, 255))
-
-                        img_textlines = img_textlines.astype(np.uint8)
-                        imgray = cv2.cvtColor(img_textlines, cv2.COLOR_BGR2GRAY)
-                        ret, thresh = cv2.threshold(imgray, 0, 255, 0)
-
-                        contours_combined, hierachy = cv2.findContours(thresh, cv2.RETR_TREE, cv2.CHAIN_APPROX_SIMPLE)
-
-                        # print(len(contours_combined),'len textlines mixed')
-                        areas_cnt_text = np.array([cv2.contourArea(contours_combined[j]) for j in range(len(contours_combined))])
-
-                        contours_biggest = contours_combined[np.argmax(areas_cnt_text)]
-
-                        # print(np.shape(contours_biggest))
-                        # print(contours_biggest[:])
-                        contours_biggest[:, 0, 0] = contours_biggest[:, 0, 0]  # -all_box_coord[int(region_final)][2]
-                        contours_biggest[:, 0, 1] = contours_biggest[:, 0, 1]  # -all_box_coord[int(region_final)][0]
-
-                        ##contours_biggest=contours_biggest.reshape(np.shape(contours_biggest)[0],np.shape(contours_biggest)[2])
-                        all_found_texline_polygons[int(region_final)][arg_min] = contours_biggest
-                        # all_found_texline_polygons[int(region_final)][arg_min]=contours_biggest
-
-                    except:
-                        pass
-                else:
-                    pass
-
-                ##cx_t,cy_t ,_, _, _ ,_,_= find_new_features_of_contoures(all_found_texline_polygons[int(region_final)])
-                ###print(all_box_coord[j_cont])
-                ###print(cx_t)
-                ###print(cy_t)
-                ###print(cx_d[i_drop])
-                ###print(cy_d[i_drop])
-                ##y_lines=all_box_coord[int(region_final)][0]+np.array(cy_t)
-
-                ##y_lines[y_lines<y_min_d[i_drop]]=0
-                ###print(y_lines)
-
-                ##arg_min=np.argmin(np.abs(y_lines-y_min_d[i_drop])  )
-                ###print(arg_min)
-
-                ##cnt_nearest=np.copy(all_found_texline_polygons[int(region_final)][arg_min])
-                ##cnt_nearest[:,0,0]=all_found_texline_polygons[int(region_final)][arg_min][:,0,0]#+all_box_coord[int(region_final)][2]
-                ##cnt_nearest[:,0,1]=all_found_texline_polygons[int(region_final)][arg_min][:,0,1]#+all_box_coord[int(region_final)][0]
-
-                ##img_textlines=np.zeros((text_regions_p.shape[0],text_regions_p.shape[1],3))
-                ##img_textlines=cv2.fillPoly(img_textlines,pts=[cnt_nearest],color=(255,255,255))
-                ##img_textlines=cv2.fillPoly(img_textlines,pts=[polygons_of_drop_capitals[i_drop] ],color=(255,255,255))
-
-                ##img_textlines=img_textlines.astype(np.uint8)
-
-                ##plt.imshow(img_textlines)
-                ##plt.show()
-                ##imgray = cv2.cvtColor(img_textlines, cv2.COLOR_BGR2GRAY)
-                ##ret, thresh = cv2.threshold(imgray, 0, 255, 0)
-
-                ##contours_combined,hierachy=cv2.findContours(thresh,cv2.RETR_TREE,cv2.CHAIN_APPROX_SIMPLE)
-
-                ##print(len(contours_combined),'len textlines mixed')
-                ##areas_cnt_text=np.array([cv2.contourArea(contours_combined[j]) for j in range(len(contours_combined))])
-
-                ##contours_biggest=contours_combined[np.argmax(areas_cnt_text)]
-
-                ###print(np.shape(contours_biggest))
-                ###print(contours_biggest[:])
-                ##contours_biggest[:,0,0]=contours_biggest[:,0,0]#-all_box_coord[int(region_final)][2]
-                ##contours_biggest[:,0,1]=contours_biggest[:,0,1]#-all_box_coord[int(region_final)][0]
-
-                ##contours_biggest=contours_biggest.reshape(np.shape(contours_biggest)[0],np.shape(contours_biggest)[2])
-                ##all_found_texline_polygons[int(region_final)][arg_min]=contours_biggest
-
-            else:
-                if len(region_with_intersected_drop) > 1:
-                    sum_pixels_of_intersection = []
-                    for i in range(len(region_with_intersected_drop)):
-                        # print((region_with_intersected_drop[i]*3+1))
-                        sum_pixels_of_intersection.append(((img_con_all_copy[:, :, 0] == (region_with_intersected_drop[i] * 3 + 1)) * 1).sum())
-                    # print(sum_pixels_of_intersection)
-                    region_final = region_with_intersected_drop[np.argmax(sum_pixels_of_intersection)] - 1
-
-                    # print(region_final,'region_final')
-                    # cx_t,cy_t ,_, _, _ ,_,_= find_new_features_of_contoures(all_found_texline_polygons[int(region_final)])
-                    try:
-                        cx_t, cy_t, _, _, _, _, _ = find_new_features_of_contoures(all_found_texline_polygons[int(region_final)])
-                        # print(all_box_coord[j_cont])
-                        # print(cx_t)
-                        # print(cy_t)
-                        # print(cx_d[i_drop])
-                        # print(cy_d[i_drop])
-                        y_lines = all_box_coord[int(region_final)][0] + np.array(cy_t)
-
-                        # print(y_lines)
-
-                        y_lines[y_lines < y_min_d[i_drop]] = 0
-                        # print(y_lines)
-
-                        arg_min = np.argmin(np.abs(y_lines - y_min_d[i_drop]))
-                        # print(arg_min)
-
-                        cnt_nearest = np.copy(all_found_texline_polygons[int(region_final)][arg_min])
-                        cnt_nearest[:, 0] = all_found_texline_polygons[int(region_final)][arg_min][:, 0] + all_box_coord[int(region_final)][2]
-                        cnt_nearest[:, 1] = all_found_texline_polygons[int(region_final)][arg_min][:, 1] + all_box_coord[int(region_final)][0]
-
-                        img_textlines = np.zeros((text_regions_p.shape[0], text_regions_p.shape[1], 3))
-                        img_textlines = cv2.fillPoly(img_textlines, pts=[cnt_nearest], color=(255, 255, 255))
-                        img_textlines = cv2.fillPoly(img_textlines, pts=[polygons_of_drop_capitals[i_drop]], color=(255, 255, 255))
-
-                        img_textlines = img_textlines.astype(np.uint8)
-                        imgray = cv2.cvtColor(img_textlines, cv2.COLOR_BGR2GRAY)
-                        ret, thresh = cv2.threshold(imgray, 0, 255, 0)
-
-                        contours_combined, hierachy = cv2.findContours(thresh, cv2.RETR_TREE, cv2.CHAIN_APPROX_SIMPLE)
-
-                        # print(len(contours_combined),'len textlines mixed')
-                        areas_cnt_text = np.array([cv2.contourArea(contours_combined[j]) for j in range(len(contours_combined))])
-
-                        contours_biggest = contours_combined[np.argmax(areas_cnt_text)]
-
-                        # print(np.shape(contours_biggest))
-                        # print(contours_biggest[:])
-                        contours_biggest[:, 0, 0] = contours_biggest[:, 0, 0] - all_box_coord[int(region_final)][2]
-                        contours_biggest[:, 0, 1] = contours_biggest[:, 0, 1] - all_box_coord[int(region_final)][0]
-
-                        contours_biggest = contours_biggest.reshape(np.shape(contours_biggest)[0], np.shape(contours_biggest)[2])
-
-                        all_found_texline_polygons[int(region_final)][arg_min] = contours_biggest
-
-                    except:
-                        # print('gordun1')
-                        pass
-                elif len(region_with_intersected_drop) == 1:
-                    region_final = region_with_intersected_drop[0] - 1
-
-                    # areas_main=np.array([cv2.contourArea(all_found_texline_polygons[int(region_final)][0][j] ) for j in range(len(all_found_texline_polygons[int(region_final)]))])
-
-                    # cx_t,cy_t ,_, _, _ ,_,_= find_new_features_of_contoures(all_found_texline_polygons[int(region_final)])
-
-                    # print(cx_t,'print')
-                    try:
-                        # print(all_found_texline_polygons[j_cont][0])
-                        cx_t, cy_t, _, _, _, _, _ = find_new_features_of_contoures(all_found_texline_polygons[int(region_final)])
-                        # print(all_box_coord[j_cont])
-                        # print(cx_t)
-                        # print(cy_t)
-                        # print(cx_d[i_drop])
-                        # print(cy_d[i_drop])
-                        y_lines = all_box_coord[int(region_final)][0] + np.array(cy_t)
-
-                        y_lines[y_lines < y_min_d[i_drop]] = 0
-                        # print(y_lines)
-
-                        arg_min = np.argmin(np.abs(y_lines - y_min_d[i_drop]))
-                        # print(arg_min)
-
-                        cnt_nearest = np.copy(all_found_texline_polygons[int(region_final)][arg_min])
-                        cnt_nearest[:, 0] = all_found_texline_polygons[int(region_final)][arg_min][:, 0] + all_box_coord[int(region_final)][2]
-                        cnt_nearest[:, 1] = all_found_texline_polygons[int(region_final)][arg_min][:, 1] + all_box_coord[int(region_final)][0]
-
-                        img_textlines = np.zeros((text_regions_p.shape[0], text_regions_p.shape[1], 3))
-                        img_textlines = cv2.fillPoly(img_textlines, pts=[cnt_nearest], color=(255, 255, 255))
-                        img_textlines = cv2.fillPoly(img_textlines, pts=[polygons_of_drop_capitals[i_drop]], color=(255, 255, 255))
-
-                        img_textlines = img_textlines.astype(np.uint8)
-                        imgray = cv2.cvtColor(img_textlines, cv2.COLOR_BGR2GRAY)
-                        ret, thresh = cv2.threshold(imgray, 0, 255, 0)
-
-                        contours_combined, hierachy = cv2.findContours(thresh, cv2.RETR_TREE, cv2.CHAIN_APPROX_SIMPLE)
-
-                        # print(len(contours_combined),'len textlines mixed')
-                        areas_cnt_text = np.array([cv2.contourArea(contours_combined[j]) for j in range(len(contours_combined))])
-
-                        contours_biggest = contours_combined[np.argmax(areas_cnt_text)]
-
-                        # print(np.shape(contours_biggest))
-                        # print(contours_biggest[:])
-                        contours_biggest[:, 0, 0] = contours_biggest[:, 0, 0] - all_box_coord[int(region_final)][2]
-                        contours_biggest[:, 0, 1] = contours_biggest[:, 0, 1] - all_box_coord[int(region_final)][0]
-
-                        contours_biggest = contours_biggest.reshape(np.shape(contours_biggest)[0], np.shape(contours_biggest)[2])
-                        all_found_texline_polygons[int(region_final)][arg_min] = contours_biggest
-                        # all_found_texline_polygons[int(region_final)][arg_min]=contours_biggest
-
-                    except:
-                        pass
-                else:
-                    pass
-
-        #####for i_drop in range(len(polygons_of_drop_capitals)):
-        #####for j_cont in range(len(contours_only_text_parent)):
-        #####img_con=np.zeros((text_regions_p.shape[0],text_regions_p.shape[1],3))
-        #####img_con=cv2.fillPoly(img_con,pts=[polygons_of_drop_capitals[i_drop] ],color=(255,255,255))
-        #####img_con=cv2.fillPoly(img_con,pts=[contours_only_text_parent[j_cont]],color=(255,255,255))
-
-        #####img_con=img_con.astype(np.uint8)
-        ######imgray = cv2.cvtColor(img_con, cv2.COLOR_BGR2GRAY)
-        ######ret, thresh = cv2.threshold(imgray, 0, 255, 0)
-
-        ######contours_new,hierachy=cv2.findContours(thresh,cv2.RETR_TREE,cv2.CHAIN_APPROX_SIMPLE)
-
-        #####contours_new,hir_new=return_contours_of_image(img_con)
-        #####contours_new_parent=return_parent_contours( contours_new,hir_new)
-        ######plt.imshow(img_con)
-        ######plt.show()
-        #####try:
-        #####if len(contours_new_parent)==1:
-        ######print(all_found_texline_polygons[j_cont][0])
-        #####cx_t,cy_t ,_, _, _ ,_,_= find_new_features_of_contoures(all_found_texline_polygons[j_cont])
-        ######print(all_box_coord[j_cont])
-        ######print(cx_t)
-        ######print(cy_t)
-        ######print(cx_d[i_drop])
-        ######print(cy_d[i_drop])
-        #####y_lines=all_box_coord[j_cont][0]+np.array(cy_t)
-
-        ######print(y_lines)
-
-        #####arg_min=np.argmin(np.abs(y_lines-y_min_d[i_drop])  )
-        ######print(arg_min)
-
-        #####cnt_nearest=np.copy(all_found_texline_polygons[j_cont][arg_min])
-        #####cnt_nearest[:,0]=all_found_texline_polygons[j_cont][arg_min][:,0]+all_box_coord[j_cont][2]
-        #####cnt_nearest[:,1]=all_found_texline_polygons[j_cont][arg_min][:,1]+all_box_coord[j_cont][0]
-
-        #####img_textlines=np.zeros((text_regions_p.shape[0],text_regions_p.shape[1],3))
-        #####img_textlines=cv2.fillPoly(img_textlines,pts=[cnt_nearest],color=(255,255,255))
-        #####img_textlines=cv2.fillPoly(img_textlines,pts=[polygons_of_drop_capitals[i_drop] ],color=(255,255,255))
-
-        #####img_textlines=img_textlines.astype(np.uint8)
-        #####imgray = cv2.cvtColor(img_textlines, cv2.COLOR_BGR2GRAY)
-        #####ret, thresh = cv2.threshold(imgray, 0, 255, 0)
-
-        #####contours_combined,hierachy=cv2.findContours(thresh,cv2.RETR_TREE,cv2.CHAIN_APPROX_SIMPLE)
-
-        #####areas_cnt_text=np.array([cv2.contourArea(contours_combined[j]) for j in range(len(contours_combined))])
-
-        #####contours_biggest=contours_combined[np.argmax(areas_cnt_text)]
-
-        ######print(np.shape(contours_biggest))
-        ######print(contours_biggest[:])
-        #####contours_biggest[:,0,0]=contours_biggest[:,0,0]-all_box_coord[j_cont][2]
-        #####contours_biggest[:,0,1]=contours_biggest[:,0,1]-all_box_coord[j_cont][0]
-
-        #####all_found_texline_polygons[j_cont][arg_min]=contours_biggest
-        ######print(contours_biggest)
-        ######plt.imshow(img_textlines[:,:,0])
-        ######plt.show()
-        #####else:
-        #####pass
-        #####except:
-        #####pass
-        return all_found_texline_polygons
 
     def save_plot_of_layout_main(self, text_regions_p, image_page):
         values = np.unique(text_regions_p[:, :])
@@ -4654,7 +4213,7 @@ class eynollah:
                     polygons_of_drop_capitals = return_contours_of_interested_region_by_min_size(text_regions_p, pixel_img)
                     # polygons_of_drop_capitals=[]
 
-                    all_found_texline_polygons = self.adhere_drop_capital_region_into_cprresponding_textline(text_regions_p, polygons_of_drop_capitals, contours_only_text_parent, contours_only_text_parent_h, all_box_coord, all_box_coord_h, all_found_texline_polygons, all_found_texline_polygons_h)
+                    all_found_texline_polygons = adhere_drop_capital_region_into_cprresponding_textline(text_regions_p, polygons_of_drop_capitals, contours_only_text_parent, contours_only_text_parent_h, all_box_coord, all_box_coord_h, all_found_texline_polygons, all_found_texline_polygons_h, kernel=self.kernel, curved_line=self.curved_line)
 
                     # print(len(contours_only_text_parent_h),len(contours_only_text_parent_h_d_ordered),'contours_only_text_parent_h')
                     pixel_lines = 6
