+++ sbb_newspapers_org_image/utils/is_nan.py
@@ -1,0 +1,3 @@
+
+def isNaN(num):
+    return num != num
